+++ src/app.module.ts
@@ -7,6 +7,7 @@ import { AuthModule } from './auth/auth.module';
 import { authConfig } from './config/auth.config';
 import { OffersModule } from './offers/offers.module';
 import { PrismaModule } from './prisma/prisma.module';
+import { reportConfig } from './config/report.config';
 import { UsersModule } from './users/users.module';
 import { VotesModule } from './votes/votes.module';
 
@@ -14,7 +15,7 @@ import { VotesModule } from './votes/votes.module';
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
-      load: [authConfig],
+      load: [authConfig, reportConfig],
     }),
     ThrottlerModule.forRoot({
       throttlers: [{ ttl: 60_000, limit: 100 }],