+++ src/app.module.ts
@@ -8,6 +8,7 @@ import { authConfig } from './config/auth.config';
 import { OffersModule } from './offers/offers.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { reportConfig } from './config/report.config';
+import { ReportsModule } from './reports/reports.module';
 import { UsersModule } from './users/users.module';
 import { VotesModule } from './votes/votes.module';
 
@@ -26,6 +27,7 @@ import { VotesModule } from './votes/votes.module';
     AuthModule,
     OffersModule,
     VotesModule,
+    ReportsModule,
   ],
   controllers: [AppController],
   providers: [