+++ src/reports/types/user-report-response.type.ts
@@ -1,0 +1,5 @@
+import { ReportReason } from '@prisma/client';
+
+export type UserReportResponse = {
+  reason: ReportReason | null;
+};